feat(notebook): extract settings into separate window
Move the settings panel out of the notebook toolbar into a dedicated settings window. Settings are now accessed via App menu > Settings... (Cmd+,) instead of a collapsible panel in the toolbar.

This clarifies that settings are global app preferences, not notebook-specific configurations. The Python/Deno dependencies button remains in the toolbar since it manages notebook-level dependencies.

Changes:
- Create new settings window with theme, runtime, and Python environment controls
- Add Cmd+, menu shortcut for opening settings (macOS convention)
- Remove collapsible settings panel from NotebookToolbar
- Remove settings gear icon from toolbar
- Use singleton pattern to prevent multiple settings windows
